This project’s objective is to program a functional and portable graphing calculator which will be able to make 2-D and 3-D graphs. The purpose of this calculator is to help students in math classes to visualize complex geometries, especially in Calculus.
The basic functioning principle of this Python graphing calculator is functional programming. As previously said, each function will compute the values for a specific mathematical function (Ex: logarithm, cosine, polynomial, etc.), the parameters for each function, like degree and coefficients for example. Each function computes all the values for the given mathematical function in a user defined domain and stores the x and y values corresponding to every point of the mathematical function’s graph into the ‘x_val’ and ‘y_val’ lists respectively and in an organized manner. Then, a library called ‘Matplotlib’ is used to plot every point into a graph. This library has a function called ‘plot(x,y)’ which takes as its arguments the x and y values and plots the corresponding point. In this program, instead of giving the function a single pair of values, we give a pair of lists which correspond to the x and y values (plot(x_val, y_val)). Then this function plots every point- in order of their index- which is contained between the two lists. Finally, a function called ‘show()’ displays the graph as a window which contains an image with the plot area and the corresponding graph(s).